[feat]: 사다리 정보를 전달하는 LadderInputDto 생성
